API

Modules

UltraNest.UltraNest

Types and constants

UltraNest.ultranest

Functions and macros

Documentation

UltraNest.UltraNest — Module

UltraNest

Julia wrapper for Python nested sampling package UltraNest.

UltraNest.ultranest — Constant

const ultranest

The Python ultranest module.

Example:

using UltraNest
smplr = ultranest.ReactiveNestedSampler(paramnames, my_likelihood, kwargs...)
result = smplr.run()

See the UltraNest Python documentation regarding usage.

Note

Convention for matrices holding multiple parameter vectors (resp. multiple samples): In UltraNest.jl, using Julia's column-major array indexing, parameter vectors are stored as rows (not columns) in the matrices.

Note

Python code must not be run on multiple Julia threads simultaneously. Either interact with ultranest only from a single Julia thread, or use the locking mechanisms provided by the PythonCall.GIL module.